Refactor monster loading to single implementation
