+++ backend/src/monster_rpg/monsters/__init__.py
@@ -6,8 +6,7 @@ from .monster_class import (
     GROWTH_TYPE_EARLY,
     GROWTH_TYPE_LATE,
 )
-from .monster_data import ALL_MONSTERS
-from .monster_loader import load_monsters
+from .monster_data import ALL_MONSTERS, load_monsters
 from .synthesis_rules import (
     SYNTHESIS_RECIPES,
     SYNTHESIS_ITEMS_REQUIRED,
