+++ backend/src/monster_rpg/monsters/monster_loader.py
@@ -1,110 +1,9 @@
-import json
-import os
 from typing import Dict, Tuple
 
-from .monster_class import Monster, GROWTH_TYPE_AVERAGE, RANK_D
-from ..skills.skills import ALL_SKILLS
-from ..skills.skill_sets import ALL_SKILL_SETS
-from ..items.item_data import ALL_ITEMS
-from ..items.equipment import ALL_EQUIPMENT
-from .monster_data import MonsterBookEntry
+from .monster_class import Monster
+from .monster_data import MonsterBookEntry, load_monsters as _load_from_data
 
 
 def load_monsters(filepath: str | None = None) -> Tuple[Dict[str, Monster], Dict[str, MonsterBookEntry]]:
     """Load monsters and monster book entries from a JSON file."""
-    if filepath is None:
-        filepath = os.path.join(os.path.dirname(__file__), "monsters.json")
-
-    try:
-        with open(filepath, encoding="utf-8") as f:
-            text = f.read().replace("\u00a0", " ")
-    except FileNotFoundError as e:
-        raise ValueError(f"Monster data file not found: {filepath}") from e
-
-    try:
-        data = json.loads(text)
-    except json.JSONDecodeError as e:
-        raise ValueError(f"Invalid monster data JSON in {filepath}") from e
-
-    monsters: Dict[str, Monster] = {}
-    book_entries: Dict[str, MonsterBookEntry] = {}
-
-    for monster_id, attrs in data.items():
-        stats = attrs.get("stats", {})
-        m = Monster(
-            name=attrs.get("name", monster_id),
-            hp=stats.get("hp", 10),
-            attack=stats.get("attack", 5),
-            defense=stats.get("defense", 5),
-            mp=stats.get("mp", 30),
-            level=attrs.get("level", 1),
-            element=attrs.get("element"),
-            speed=stats.get("speed", 5),
-            ai_role=attrs.get("ai_role", "attacker"),
-            growth_type=attrs.get("growth_type", GROWTH_TYPE_AVERAGE),
-            monster_id=monster_id,
-            rank=attrs.get("rank", RANK_D),
-            image_filename=attrs.get("image_filename"),
-        )
-        # skills
-        skill_ids = list(attrs.get("skills", []))
-        # include skills from referenced sets
-        for set_id in attrs.get("skill_sets", []):
-            sset = ALL_SKILL_SETS.get(set_id)
-            if not sset:
-                continue
-            for lvl, ids in sset.get("learnset", {}).items():
-                if lvl <= m.level:
-                    if isinstance(ids, str):
-                        ids = [ids]
-                    for sid in ids:
-                        if sid not in skill_ids:
-                            skill_ids.append(sid)
-        # additional skills
-        for sid in attrs.get("additional_skills", []):
-            if sid not in skill_ids:
-                skill_ids.append(sid)
-
-        skills = [ALL_SKILLS[s] for s in skill_ids if s in ALL_SKILLS]
-        m.skills = skills
-        # drop items
-        drops = []
-        for item_id, rate in attrs.get("drop_items", []):
-            item = ALL_ITEMS.get(item_id) or ALL_EQUIPMENT.get(item_id)
-            if item:
-                drops.append((item, rate))
-        m.drop_items = drops
-        # learnset
-        learnset = attrs.get("learnset", {}).copy()
-        for set_id in attrs.get("skill_sets", []):
-            sset = ALL_SKILL_SETS.get(set_id)
-            if not sset:
-                continue
-            for lvl, ids in sset.get("learnset", {}).items():
-                key = str(lvl)
-                existing = learnset.get(key)
-                if existing is None:
-                    learnset[key] = ids
-                    continue
-                if isinstance(existing, str):
-                    existing = [existing]
-                if isinstance(ids, str):
-                    ids = [ids]
-                for sid in ids:
-                    if sid not in existing:
-                        existing.append(sid)
-                learnset[key] = existing
-        m.learnset = {int(k): v for k, v in learnset.items()}
-
-        monsters[monster_id] = m
-
-        book = attrs.get("book", {})
-        book_entries[monster_id] = MonsterBookEntry(
-            monster_id=monster_id,
-            description=book.get("description", ""),
-            location_hint=book.get("location_hint", ""),
-            synthesis_hint=book.get("synthesis_hint", ""),
-            reward=book.get("reward", 0),
-        )
-
-    return monsters, book_entries
+    return _load_from_data(filepath)
